refactor: auth 관련 훅 리팩토링
이름 및 위치 변경 등
