+++ frontend/src/Router.tsx
@@ -1,7 +1,8 @@
 import { Layout } from 'Layout';
+import { useEffect } from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
-import useAuth from 'hooks/useAuth';
+import useAuth from 'hooks/auth/useAuth';
 
 import {
   Feed,
