+++ frontend/src/pages/ChallengeDetailPage/index.tsx
@@ -6,7 +6,7 @@ import styled, { ThemeContext } from 'styled-components';
 import { getEmoji } from 'utils/emoji';
 
 import usePostJoinChallenge from 'hooks/api/usePostJoinChallenge';
-import { useManageAccessToken } from 'hooks/useManageAccessToken';
+import useLogout from 'hooks/auth/useLogout';
 import useSnackBar from 'hooks/useSnackBar';
 
 import { ChallengeExplanationTextProps } from 'pages/ChallengeDetailPage/type';
@@ -21,7 +21,7 @@ const makeCursorPointer = {
 
 export const ChallengeDetailPage = () => {
   const renderSnackBar = useSnackBar();
-  const checkLogout = useManageAccessToken();
+  const logoutByError = useLogout();
   const navigate = useNavigate();
   const themeContext = useContext(ThemeContext);
   const { challengeId } = useParams();
@@ -31,7 +31,7 @@ export const ChallengeDetailPage = () => {
     {
       refetchOnWindowFocus: false,
       onError: (error) => {
-        if (checkLogout(error)) {
+        if (logoutByError(error)) {
           return;
         }
